Damaged Wall and Layout Updates
Changed the damaged wall from moveable to static to prevent the meshes from disappearing. Adjusted the scale to make a more cohesive layout, which caused several changes to the structure of the damaged wall. Made a few adjustments to the layout, specifically the wall to the right of the entrance (if you are standing in the main circle).